test(desktop): decouple notification pane resolution
