+++ apps/desktop/src/main/lib/notifications/resolve-pane-id.ts
@@ -1,3 +1,4 @@
+import { appState } from "../app-state";
 import type { TabsState } from "../app-state/schemas";
 
 /**
@@ -47,6 +48,36 @@ export function resolvePaneIdFromTabsState(
 				return existingPaneId;
 			}
 		}
+	}
+
+	return undefined;
+}
+
+/**
+ * Resolves pane IDs using main-process persisted tabs state when possible.
+ * Explicit pane IDs are trusted to avoid dropping early hook events while
+ * tabsState persistence is still catching up.
+ */
+export function resolvePaneId(
+	paneId: string | undefined,
+	tabId: string | undefined,
+	workspaceId: string | undefined,
+	sessionId: string | undefined,
+): string | undefined {
+	if (paneId) {
+		return paneId;
+	}
+
+	try {
+		return resolvePaneIdFromTabsState(
+			appState.data.tabsState,
+			undefined,
+			tabId,
+			workspaceId,
+			sessionId,
+		);
+	} catch {
+		// App state not initialized yet, ignore
 	}
 
 	return undefined;
